requirements.txt
# db
PyMySQL==0.9.3
PyYAML==5.3.1

# google
google-api-python-client==1.10.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
oauth2client==4.1.3

# testing tools
pytest==6.0.1
coverage==5.2.1